web: split AccountMenu into a standalone component
