+++ web/src/SidebarAccount.test.tsx
@@ -5,12 +5,7 @@ import { act } from "react-dom/test-utils"
 import ReactModal from "react-modal"
 import { MemoryRouter } from "react-router-dom"
 import ShortcutsDialog from "./ShortcutsDialog"
-import SidebarAccount, {
-  MenuContentButtonSignUp,
-  MenuContentButtonTiltCloud,
-  SidebarAccountRoot,
-  SidebarMenuContent,
-} from "./SidebarAccount"
+import SidebarAccount, { SidebarAccountRoot } from "./SidebarAccount"
 
 beforeEach(() => {
   ReactModal.setAppElement(document.body)
@@ -31,34 +26,6 @@ it("renders nothing on a Tilt Snapshot", () => {
 
   expect(root.find(SidebarAccountRoot)).toHaveLength(0)
   root.unmount()
-})
-
-it("renders Sign Up button when user is not signed in", () => {
-  const root = mount(
-    <SidebarMenuContent
-      tiltCloudUsername=""
-      tiltCloudSchemeHost="http://cloud.tilt.dev"
-      tiltCloudTeamID=""
-      tiltCloudTeamName=""
-      isSnapshot={false}
-    />
-  )
-
-  expect(root.find(MenuContentButtonSignUp)).toHaveLength(1)
-})
-
-it("renders TiltCloud button when user is signed in", () => {
-  const root = mount(
-    <SidebarMenuContent
-      tiltCloudUsername="amaia"
-      tiltCloudSchemeHost="http://cloud.tilt.dev"
-      tiltCloudTeamID="cactus inc"
-      tiltCloudTeamName=""
-      isSnapshot={false}
-    />
-  )
-
-  expect(root.find(MenuContentButtonTiltCloud)).toHaveLength(1)
 })
 
 it("renders shortcuts dialog on ?", () => {
